Remove last uses of "possibly-insecure" mktemp(3)
The GNU linker (and perhaps others?) has complained about uses of
mktemp(3) being "potentially insecure" for a long time (RT #6199
also noted this).  Our usage of the function has decreased with time,
but has not disappeared entirely.

kprop's credentials never need to hit the disk, so a MEMORY keytab
is sufficient (and does not need randomization).
store_master_key_list is explicitly putting keys on disk so as to
do an atomic rename of the stash file, but since the stash file
should be in a root-only directory, we can just use a fixed name
for the temporary file.
When using this fixed name, we must detect (and error out) if the
temporary file already exists; add a test to confirm that we do so.

ticket: 1794