Clean up stash file error handling
The comment previously failed to match the behavior; the intent is that
if we failed to write out the entire stash file into the temporary
location, we should remove the partial file.
However, we can just always try to remove the partial file, and not
worry about whether it already exists.