Cleanup ksu.h
Move declarations of (most) functions only used in one compilation unit
to that source file and mark as static.  Remove the 'krb5_' prefix from
such functions, if applicable.

Move declarations to the top of the source file (before comments
describing the first function defined in the file), and other minor
whitespace fixes.