Use a real client caller for schpw requests
Previously, when a principal changed its password using 'kpasswd'
or has its password set using 'ksetpwd', the database entry would
show up as last modified by 'kadmin@REALM', instead of the actual
principal which authenticated to make the change.
This is because the kpasswd requests go through a separate daemon
which has a connection open to the database acting as kadmind, and
prints tickets for kadmin/changepw to verify incoming requests.
The server handle is still marked as open by kadmind, though, which
is what is recorded in the database when the entry is changed.
Work around this cosmetic defect by forcing the current caller field
to be the actual client for just this request.