Use default alignment on OS X
For historical compatibility, we force 16-bit alignment on the mac
platform.  The SAP GUI Java application expects the default
alignment for GSS data structures, and experiences segfaults on
64-bit architectures when interacting with a 16-bit aligned GSS-API
library.

(gss_buffer_t is a 32-bit length followed by a void *.  On 32-bit
machines, the void * is properly aligned immediately after the length
field, so everything works fine.  On 64-bit machines the pointer
is 64-bits and as such defaults to 64-bit alignment, leaving 32-bits
of padding between the length and the pointer, which is a different
binary format.)