Tweak error handling a bit
Just respond if we can't initialize the audit state.  Even setting
a status string and jumping to the end would not help here, as
finish_process_as_req() would try to dereference the (NULL) au_state.

Also report failure of kau_make_tkt_id() when generating the output
ticket ID.  A failure here is still a failure of the KDC routine, even
though a ticket is constructed and ready to return to the client;
we would be violating our audit contract to return successfully here.