Hack in a windows ribbon
None of this MFC nonsense.

We have new classes to implement IUIApplication and IUICommandHandler,
so give the MainFrame a UIApplication member variable and have
the main frame call out to initialize all the UI bits, including the
ribbon, during startup.

Build in the new objects, and things seem to work as-expected.
(The main frame doesn't know to subtract the height of the ribbon,
but we don't expect that to work, yet.)