Make ribbon home tab more pretty
In the absence of a label title for a given command group, the
text "MS Shell Dlg" was inserted, which is somewhat confusing to
the user.  Give the ticket-maniuplation group a title consisting
of a space, which removes the confusing text from the display.

ticket: 8206 (new)
queue: kfw
tags: pullup
target_version: 1.13.3