Let the MSLSA: cache respect [capths]
The get_cached_tgt() helper will cause the LSA to make a TGS-REQ
if there is no appropriate TGT in the cache, due to the design
constraints of the MSLSA: cache type.  The LSA appears to always
attempt a direct cross-realm relationship in non-AD environments,
which ignores the [capths] configured in krb5.ini.  As such, the
call to cache_get() can return KRB5KDC_ERR_S_PRINCIPAL_UNKNOWN
from the LSA, which should be treated the same as KRB5_CC_NOTFOUND
for the case of retrieving a cross-realm TGT.

This will allow later logic which follows the capths to retrieve
an appropriate cross-realm TGT to proceed.